+++ src/com/intellij/idea/plugin/hybris/common/HybrisConstants.kt
@@ -209,6 +209,8 @@ object HybrisConstants {
     const val ANT_HEAP_SIZE_MB = 512
     const val ANT_STACK_SIZE_MB = 128
 
+    const val BS_TYPE_OBJECT = "java.lang.Object"
+
     const val TS_MAX_RECURSION_LEVEL = 2
     const val TS_TYPE_OBJECT = "java.lang.Object"
     const val TS_RELATION_ORDERING_POSTFIX = "POS"